requirements/test_requirements.txt
pytest==3.7.0
pytest-cov==2.5.1
pytest-django==3.1.2
codacy-coverage==1.3.6
selenium==3.14.0
factory_boy==2.11.1